HACK: ARM: OMAP2+: prevent pm_qos going overboard
The computation of various collated bus requests to actual L3 frequency
is not really consistent across all SoCs. Unfortunately, the current
code is based from OMAP3 days, which is not accurate for OMAP4 or
OMAP5 SoC family. In short, on later SoCs, we are probably
under-estimating the bandwidth to frequency map. This results
in frequency request to DVFS layer which cannot be supported for
L3 device.

While the final solution is being finalized, hack to allow all
requests to pass even if computation shows overboard opp frequency.

Hack is to check for a floor frequency (lower than estimated frequency)
if a ceil frequency is not available.

NOTE: To ensure that we fix this eventually, an ratelimited debug
is fired out when the match fails - this should be converted to rate
limited error once we have appropriate computational fix.

Change-Id: Id0726c10ef7f752f25b8ac34da79dabefd495d7f
Signed-off-by: Nishanth Menon <nm@ti.com>