USB: DWC3: prevent system low power state when device is enumerated
The suspend hooks are now added to prevent dwc3 from entering low power state
whenever usb is connected.

This is necessary because when the dwc3 module is disabled, the phy and dwc3
goes out of sync and both dwc3 and phy has to be reset to make it functional.
This will make the device to be disconnected and re-enumerated again. To avoid
device disconnect, dwc3 is prevented from entering low power state.

Change-Id: I285f37faa4ca328934909ff8a03f503e6a5f6e27
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>