PM QoS: Add debugfs support to view the list of constraints
PM QoS requests are notoriously hard to debug and made even
more so due to their higly dynamic nature. Having visibility
into the internal data representation per constraint allows
us to have much better appreciation of potential issues or
bad usage by drivers in the system.

So introduce for all classes of PM QoS, an entry in
/sys/kernel/debug/pm_qos that shall show all the current
requests as well as the snapshot of the value these requests
boil down to. For example:
==> /sys/kernel/debug/pm_qos/cpu_dma_latency <==
1: 4444: Active
2: 2000000000: Default
3: 2000000000: Default
4: 2000000000: Default
Type=Minimum, Value=4444, Requests: active=1 / total=4

==> /sys/kernel/debug/pm_qos/memory_throughput <==
Empty!

==> /sys/kernel/debug/pm_qos/network_latency <==
Empty!

==> /sys/kernel/debug/pm_qos/network_throughput <==
Empty!

The actual value listed will have their meaning based
on the QoS it is on, the 'Type' indicates what logic
it would use to collate the information - Minimum,
Maximum or Additive. Value is the collation of all
requests. This interface also compares the values with
the defaults for the QoS class and marks the ones that
are currently active.

Change-Id: Ib393517c674f144a48b9964539e6529bc3741f69
Signed-off-by: Nishanth Menon <nm@ti.com>