USB: DWC3: add PM support for dwc driver
Implements PM wherein the module gets enabled and initialized only
when the cable is connected and a gadget driver (e.g. g_zero) is inserted. And
the module is disabled when the cable is disconnected or gadget driver is
removed.

once the dwc3 module gets disabled, the synchronization between phy and dwc3
goes bad and so the phy and dwc3 core has to be reset again in order to get it
functional. This resetting is done during every connect.

[m-sonasath@ti.com] Fixed the dwc->is_connected logic in the patch and logically
sequenced the flow within dwc_core_late_init().

Change-Id: I78d415fea2d1e2a6a18f3fda28fb30d8e2298da0
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Partha Basak <p-basak2@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>