HACK: OMAP: rpmsg: do a frequency search on floor for ducati requests
Ducati seems to have not considered SoC's current frequency
capability which may not support high frequencies required
for some use cases. As a result, the requests will fail when
attempted on OMAP PM framework as they are out of supported
range.

With this hack, we essentially run the system at lower than
requested frequency when the requests are over the maximum
enabled frequencies for that device.  As a result, the usecase
may or may not work due to the lack of sufficient frequency.

The alternative of this hack is to never run those usecases.
We, for the moment, live with the "may or may not work", until
we pass the current SoC configuration to Ducati and Ducati uses
this information to request a frequency with better judgement.

REVERT THIS PATCH ONCE DUCATI-SYSLINK GET BETTER HANDLE on
SYSTEM's AVAILABLE RESOURCES.

Change-Id: I89679ab5cc5588d9202708229b565a0751029258
Signed-off-by: Chandra Sekhar.Anagani <chandu@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>