OMAPDSS: HDMI: Wait for audio to complete before turning off PHY power
When HDMI is disconnected HDMI phy power needs to be turned off after
audio that is in the process of rendering completes. Wait for the
HDMI_WP_AUDIO_CTRL[31] bit to reset before turning phy power off.

In the event of a mode change, i.e. power off->on no waiting is done.

Change-Id: Ieb2a88e3a7e048c0ab918fbefea322769f415eb6
Signed-off-by: Priyesh Bisla <a0271372@ti.com>