USB: DWC: dwc3-omap: Fix enum when USB cable is connected during suspend state
If a USB cable is connected when the system is suspended, the enumeration fails
since the get_sync in dwc fails. This is root caused to get_sync being called
even before the runtime is fully up. It is fixed by scheduling a delayed work
that sets the mailbox after some delay.

[m-sonasath@ti.com]
dwc->is_connected is used to prevent the system from entering suspend if a work
has been scheduled but not yet been executed.

Change-Id: I649b3fe8a971fbe376e49b47f42ff73e97e72acf
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>
Signed-off-by: Partha Basak <p-basak2@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>