ARM: OMAP5: Reduce the static IO mapping
EMIF, GMPC and DMM driver can ioremap() the address
space as part of driver intialisation and there is
no need to have static IO mapping for them.

Hence remove the un-used static IP space and let
the respective drivers manage it as part if driver
init.

Based on the changes that was earlier done for OMAP4
which went as a part of OMAP2+ cleanup series.

Ref: http://www.spinics.net/lists/linux-omap/msg69238.html
by Santosh Shilimkar <santosh.shilimkar@ti.com>

Signed-off-by: R Sricharan <r.sricharan@ti.com>

minor:
port to p-linux-omap-3.4

Change-Id: Ic7dc2a61406513d995b9c9b8223a39a6f8e0e797
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>