caif-hsi: Add support for dynamic frequency scaling
This patch a squach + fixup of some commits

Change-Id: I41469b19b2ef91e8070147ff68d4b95229a03e29

Signed-off-by: Sjur BRENDELAND <sjur.brandeland@stericsson.com>

----
Add support for changing frequency used on the HSI interface.
A CAIF specific frequency scale request/response is used to
negotiate the speed change. Frequency changes is currently
based on a simple queue measurements, combined with a grace
period to avoid frequent occilations.

Signed-off-by: Per Ellefsen <per.ellefsen@stericsson.com>
Change-Id: I41469b19b2ef91e8070147ff68d4b95229a03e29
Reviewed-by: Sjur BRENDELAND <sjur.brandeland@stericsson.com>
Tested-by: Sjur BRENDELAND <sjur.brandeland@stericsson.com>