OMAP5:EHCI: Fix unbalanced pm calls in EHCI driver.
Its been observed in some scenarios that from the USB host stack
pm runtime managment that even if the power.usage_count and power.disable_depth
are zero the EHCI device status as active. This will prevent the accquiring of
clocks as the get_sync fails and results in crash.

Change-Id: I524dae2007b0e3ce2077e2a2d1034b47cebd606d
Signed-off-by: Mantesh Sarasetti <mantesh@ti.com>
Signed-off-by: Poulain-XID, Loic <l-poulain@ti.com>