OMAP5: I2C: Enable internal pull-ups for I2C lines
Pull-ups are required to ensure SCL & SDA lines are held
high and to ensure the lines are pulled from low to high
within the time. Based on the load external pull-ups are
connected to these lines. In case of missing external
pull-ups, it is recommended to enable internal pull-ups.

On omap5 sEVM board none of the I2C controllers have
external pull-ups connected. To prevent timeouts and
maintain line integrety this patch enables internal
pull-ups. If a line has external pull-up connected this
should be disabled.

Change-Id: I481efcab2ccd57cd6cfbb13ad6b6d70e99b60c16
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>
Signed-off-by: Ivan Evlogiev <x0099089@ti.com>
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>