omap5: pm: Add static dep on DSS towards EMIF
DSS underflow is seen when EMIF is allowed to idle while system
cpuidle is enabled, specificly when MPU is allowed to idle. The
issue is not seen in C1 state because MPU has a static dep on EMIF.
Enabling SD from DSS to EMIF is solving this issue.

REVISIT: This is a temporary WA till the right fix is identified.

Signed-off-by: Girish S G <girishsg@ti.com>
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>