omap5: bandgap: Fix gpio used in bandgap
In function omap_bandgap_platform_build, incorrect field
was used to initialise the gpio line.

The gpio line to be used for tshut is configured as
platform data in scm control module. So when the bandgap
module is getting initialised, need to use this data.

Also removed the unused tshut_gpio field from the omap_bandgap_data
structure.

Change-Id: I9212fa5c1a2e1e1ecdebc7b907181bbb1fbdda80
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>