I2C: OMAP: call pm_runtime_get_sync only if hwspinlock is locked
Currently, we do a runtime get_sync and then check for h/w spinlock.
If getting hwspinlock failed, then we return without doing a put.
This results in clock to be active on I2C module preventing low
power state in this condition.

So, optimize logic to check if lock was captured before getting
clock for the I2C module and handle error appropriately.

As part of this change, add a sanity check to ensure that
we have valid omap_i2c_dev pointer for us to operate on.

Change-Id: Ie9bfc40a08eb60024bd006286e56695c11da6f93
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>