omap: iommu: enable bus-error back response to remoteproc
The OMAP remoteproc MMUs can give out the fault address as well
as the PC value that caused the fault. This feature is not
available on ARM cores, and instead the value is retrieved
using a bus-error response back to the remote processor.

This functionality is programmable through the MMU_GP_REG
register (only defined for IPU).  The cores are simply
stalled if the MMU_GP_REG.BUS_ERR_BACK_EN bit is not set.
When set, a bus-error exception is raised allowing the
processor to handle it as a bus fault and print the
exception context.

Change-Id: I394a60b1a0cfa47b169f6b56f1efcae1045a07a9
Signed-off-by: Wajahat Khan <w-khan@ti.com>
Signed-off-by: Shahid Akhtar <sakhtar@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>