OMAPDSS: DISPC: Force L3_2 CD to NOSLEEP when dispc module is active
It has been identified that L3_2 CD is idling and not responding to the traffic
initiated by initiators. As per errata i740 worakaround, ISS-L3_2 noidle
constraint should be in place when ISS is effective.

But random hangs are observed if DSS-L3_2 is not enabled, and this patch is
a temporary workaround till actual issue is rootcaused and fixed.

Change-Id: I423dcc4a108929ace1cdf6e70ab94c098509d47b
Signed-off-by: Devaraj Rangasamy <dev@ti.com>