omap5: bandgap: debug error print
On Omap5 there are 3 bandgap sensors, cpu, gpu and core.
And for cpu and gpu domains, there are governor and cooling
devices identified. But for core domain, currently there are
no governor nor cooling devices. When the core temperature sensor
updates the temperature report for this core domain, thermal framework
returns an error. Since this temperature update happens quite often,
there are many repetitive error messages due to this error debug print.
Hence making this debug statement as pr_err_once, and once the governor and
cooling devices are identified for core, can revert it back to dev_err.

Change-Id: I807cce8817b2a5f5875185c1d1324779687e6ec4
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>