OMAP5: WLAN: Fix mux settings for WLAN GPIO
The present mux configuration sets the pin as only o/p pin.
It needs pull-up to be enabled so that the line is not pulled
low during OFF due to GPIO pin pulling low caused due to HW
issue on GPIO o/p pins (BUG ID: OMAP5430-1.0BUG0166). This
GPIO issue is on ES 1.0 and will be fixed on ES 2.0. This patch
can then be reverted.

Change-Id: I18fae33ee94d883aa3634191a050dfaf051fb648
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>