omap4plus_min_defconfig: simplify by disable part twl modules
As TWL6035 RTC seems to have issues on certain early samples
with system freeze, disable it in pm_defconfig for suspend/resume to work
reliably.

Disable CONFIG_TWL6040_CORE. This should be enabled by
default in Kconfig when sound driver is properly adapted with
PM framework.

We do leave on rest of TWL6030 core etc as we still need regulator
framework and other features to continue to function.

Signed-off-by: Girish S G <girishsg@ti.com>