TI thermal framework: added "case" thermal governor
This patch adds the "case" thermal governor into the thermal framework.
The "case" temperature sensor will report temperature to the "case"
thermal governor that could throttle any cooling device registered
on this domain, like CPU frequency, brightness, charging, etc.

This governor exports its thot and tcold thresholds on
the thermal framework debug support. This can be used
on calibration process.

The used cooling level is starting from 0. The cooling level
is assumed to be a representation of cooling devices priorities.
Once the governor sees that we enter the alert zone (between
tcold and thot), it will start to broadcast cooling messages
to its cooling devices, which should take the mitigation
action accordingly. It is not required to take an action
on every cooling message though. This set of actions,
priorities and order must be defined by system designer
and integrator.

In other words, once into alert zone, it will increase
the cooling level by 1 on each subzone inside the alert
zone. The alert zone has been split into several subzones,
each representing a cooling priority that needs to be taken.
The governor will ask for cooling agent to act on each subzone,
depending on their priorities. They must interpret the cooling_level
as a cooling priority request.

This governor exports the amount of actions that are not done
before the critical point is reached. This way the cooling
devices can decide which level it should start off the aggressive
actions.

As a preventive action, this governor performs a orderly_poweroff
if the maximum cooling level is reached. This means that if all
programmed actions failed to cool down the system, the governor
will notify userland and shutdown.

Change-Id: I05f0dc2960fbc14784974683ffe0cf8234d1808f
Signed-off-by: Sebastien Sabatier <s-sabatier1@ti.com>
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>