omap: iommu: add latency request information to device data
The iommu driver can make use of the dev_pm_qos/pm_qos apis to
ensure that latency constraints are enforced from the time that
the device is enabled to the time that it gets disabled.

This patch adds initial support in the iommu device. The required
fields are introduced in the device attribute, platform data and
iommu data structures.

Change-Id: Ia67a26a4e6367afc1f07582b6fd39f471cb1e7d5
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>