ASoC: omap-abe-twl6040: Use exact McBSP fclk rate
Clock rate used to configure McBSP functional clock was calculated
based on the stream rate. This approach fails when McBSP is connected
to an ABE BE and that BE is used by FEs with different sample rate.
In this scenario, there are two functional clock rates requested so
the last to get requested fails.

The actual clock rate of McBSP functional clock can be retrieved using
the clock framework. Specifically, the rate of the internal clock (also
known as "prcm_fck") is queried as this will be the parent of McBSP
functional clock.

Change-Id: I486664401f3f6246a86d7b9795a8938765a6d048
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>