ARM: OMAP4+: PM: De-assert CLKREQ in Device RET and OFF states
CLKREQCTRL.CLKREQ_COND is used to manage system clock gating.

And, this system clock gating control is done whatever
the source providing this clock. In both Crystal or Oscillator
configurations, system clock can be gated depending on
the CLKREQCTRL.CLKREQ_COND register value.

Moreover, it's correct that pwrreq can't toggle if system clock
is kept running. Pwrreq can be deasserted only after gating of
system clock and this is independent of sysboot configuration.

So configure CLKREQ_COND to gate clock for retention and OFF state.

Change-Id: If3c5f05c4f3ae8653fe150e370832faf0cb6fe19
Signed-off-by: Rajendra Nayak <rnayak@ti.com>
[ported from k3.0 commit:88193731]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>