ARM: OMAP: PM: program core_pd in omap_pm_begin
Set DEV OFF as part of core_pd programming in omap_pm_begin.
cpu notifier call chain gic save/restores is based on DEV
OFF status. And enabling of DEV OFF should be done
before cpu notifiers is triggered.

However, once the omap5_suspend handler is invoked, we ensure
that saved_state is ON so that on resume we restore the power
state of Core PD back to ON and continue operation.

Signed-off-by: Girish S G <girishsg@ti.com>