OMAPDSS: CEC : Fix CEC power on crash on boot up
In default hdmi display we want to pick the best timings for the
TV which is connect and allocate the framebuffer of that size.
For this to happen we read edid in omapfb which only need the
hdmi DDC clock ON. Once we have the best & biggest WXH we will
call set_timings call to driver to setup the HDMI panel. But
at this moment HDMI which is default display is still not turned
on by calling driver->enable.

But the driver->set_timing call driver->disable set the timings
and then driver->enable.

If done so driver->disable will call cec register without hdmi
clocks ON.

To avoid the crash call CEC power down only if HDMI panel
is in active state.

Change-Id: I5f994cb2dce0dd2d9b7d46a1a8f20c93eda2deb8
Signed-off-by: Dandawate Saket <dsaket@ti.com>