ARM: OMAP4+: PM: de-couple device-off state from core_pd transition
Currently there are bunch of activities that are performed as part
of cluster notifiers including potentially powerdomain sensitive
operations, GIC handling etc. Since, these operations require to
know the power state transitioning to, they use the device_off
setting to determine operations related to OFF mode need to be performed.

Unfortunately, in our current implementation, device_off is strongly
coupled with device_off configuration. This is inherently dangerous
as it splits the powerdomain programmation across multiple entry points
which can leave the system in unbalanced state on error handling paths.

Hence, decouple DEVOFF association with CORE power domain state.
DEV OFF will only be achieved when all power domains are in OFF/OSwR. So, from
the code flow it really doesn't depict the actual TRM definition.

In-order to do this, we introduce pwrdm_read_device_off_state
as an query API and  pwrdm_enable_off_mode to program the device_off
state.

The device_off is enabled and disabled at pm_begin and exit routines, while
the power domain programmation is done as part of suspend path. There is
no plans of usage of device_off as part of CPUIdle path due to the
inherent latencies involved.

Note: This partly reverts the patch by Tero Kristo <t-kristo@ti.com>
"OMAP4: pwrdm: add support for device off" as the functionality is
meaningless for OMAP4+ at this point in time, in addition to simplifying
the code flow for maintainability.

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>