Thermal: Add a PCB domain check in the governor
The governor should only check the PCB temp if one
exists.  So for every temp reported the governor will
check to see if the PCB domain exists.

This will simplify the calls to the framework due to the
governor not having to request something that does not exist
it can just skip the PCB lookup temp calls.

Change-Id: I07b05474065922becc72d5ecc35b83c258291359
Signed-off-by: Dan Murphy <dmurphy@ti.com>