ARM: OMAP5: PM: Fix HAL API Index for OMAP5 to > 2.1.0 PPA
PPA index value for OMAP5 based on PPA 2.1.0+ is now
from 0x200. We could finally be using 0x200 service
to find the index and use the index for OMAP4 and 5
in a consistent manner. for the timebeing update the
macro to stay in sync with PPA under use.

NOTE: This breaks compatibility with PPA revisions
*lower* than PPA 2.1.0.

Signed-off-by: Madhusudan C M <c-madhu@ti.com>