ARM: OMAP3+: Implement dmtimer workaround for errata i767
Errata Title:
GP and WD Timer Returns Incorrect Time Due to Dynamic Clock Gating

Description:
After wakeup from IDLE state, it is possible to read the old value of the
counter (the value before IDLE state) on general-purpose timers and watchdog
timers. Only impacts posted mode.

Workarounds:
1). Disable posted mode
2). Use static dependency between timer clock domain and MPUSS clock domain
3). Use no-idle mode when the timer is active

Workarounds #2 and #3 are not pratical from a power standpoint and so
workaround #1 has been implemented. Disabling posted mode adds some CPU overhead
for configuring the timers as the CPU has to wait for the write to complete.
However, disabling posted mode guarantees correct operation.

Please note that it is safe to use posted mode for timers if the counter
register will never be read. An example of this is the clock-event system
timer. This is used by the kernel to schedule events however, the timers
counter is never read. Given that the kernel configures this timer often yet
never reads the counter register it is safe to enable posted mode in this case.
Hence, for the timer used for kernel clock-events, posted mode is enabled by
overriding the errata for devices that are impacted by this defect.

Although both dmtimers and watchdogs are impacted by this defect this patch only
implements the workaround for the dmtimer. The workaround for the watchdog timer
has been implemented separately by patch "ARM: OMAP5: Implement workaround for
sync-timer and watchdog errata".

It should be noted that when using posted mode there is another defect that can
cause the timer counter to return an incorrect value, if the timer interface
clock is greater or equal to 4 times the function clock frequency. By disabling
posted mode this defect is also avoided.

Signed-off-by: Jon Hunter <jon-hunter@ti.com>