ARM: OMAP5: Implement workaround for sync-timer and watchdog errata
Errata Titles:
- GP and WD Timer Returns Incorrect Time Due to Dynamic Clock Gating (i767)
- 32K Timer Returns Incorrect Time Due to Dynamic Clock Gating (i768)

Errata Description (same for i767 and i768):
After wakeup from IDLE state, it is possible to read the old value of the
counter (the value before IDLE state) for the 32-kHz synchronized timer.

Errata Workaround:
1). Disable posted mode (only applicable for MPU watchdog)
2). Add a static dependency between the MPUSS and WKUPAON clock domains

This patch add a static dependency between the MPUSS and WKUPAON clock domain
to ensure that the 32k sync-timer and MPU watchdog timer returns the correct
counter value. Typically adding a static dependency is not desirable from a
power perspective as it will cause unnecessary wake-ups of a power and clock
domain. However, given that these timers are in an always-on power domain the
impact is not as significant.

Please note that i767 for the watchdog timer could also be workaround by
disabling posted for the watchdog. However, given that we need to add a static
dependency for the sync-timer this also ensures that i767 is avoid for the MPU
watchdog timer.

Signed-off-by: Jon Hunter <jon-hunter@ti.com>