ARM: OMAP: powerdomain: consider current state when not deciding to skip programing state
The current code in omap_set_pwrdm_state checks if
next_state is the same as the pwr state that it
was asked to go to. It ignores the current state
of the power domain. This results in the following
combinations:
current	next	prev	result
OFF	OFF	OFF	wont program	OK
OFF	OSWR	OSWR	wont program	NOT OK
OFF	CSWR	CSWR	wont program	NOT OK
OSWR	OFF	OFF	wont program	NOT OK
OSWR	OSWR	OSWR	wont program	OK
OSWR	CSWR	CSWR	wont program	NOT OK
CSWR	OFF	OFF	wont program	NOT OK
CSWR	OSWR	OSWR	wont program	NOT OK
CSWR	CSWR	CSWR	wont program	OK

However, by ignoring the current state, the requested
power domain will not be in the power state that
we expect it to be in. This is a violation of the contract
of omap_set_pwrdm_state that it should be able to achieve
the power domain state that was targetted to achieve

Reported-by: Mahesh Renduchintala <mahesh@ti.com>
Acked-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>