ARM: OMAP: Powerdomain: introduce power state comparison API
With various possible values of Macro definitions used
to describe power domain states, it is imperative that
comparisons between power domain states do not result
in subtle bugs in code flow path. Inorder to do this,
powerdomain maintains an internal lookup table which is
kept upto date and is used for comparison when users need
to compare powerdomain states.

In theory, if we use the current macro definition values,
we could use equations such as:
(a) ? a | (1 << 2) : a
or
(a & 0x2) ? a | (1 << 3) : a

However any future change in the macro definition will
defeat the translation logic in addition to being
a bit more obscure. we pay instead the penalty of
an ldr and a lookup table storage.

[grygorii.strashko@ti.com, c-madhu@ti.com: co-development]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Madhusudan C M <c-madhu@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>