arm: omap: devices: expand init_omap_dev function
The init_omap_dev helper function is used to construct the omap
devices required for MM usecases (like FDIF & ISS), and used by
the rpmsg resmgr layer. This function interface has been expanded
to allow device pm latencies to be passed, to allow plug-in of
custom activate and deactivate functions.

Change-Id: I37fca6010571e5040459f254f157b391c2c753d0
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>