drivers: usb: omap-usb3: Decrease the number of transitions to recovery
Power-cycling the PHY (partially) decreases the number of transitions to
"Recovery" state. Hence changed the power up sequence to a partial
power-up before a full power-up of the PHY.

HW Issue background:
According to PHY expert, the indication that the DLL part of Ganga is
not relocking is indicated by GANGARX_OBSERVE_REG3.b29.
When the clock is present and before cutting it, this bit is 1.
It transitions to 0 when the clock is cut, but does not transition to 1
again when the clock is restablished. A power-cycle of the RX-phy
(PWRCTRL_CMD[0]=0 then 1) recovers from that situation i.e. the b29
transitions to 1. A partial power-cycle of the RX-phy
(PWRCTRL_CMD[6]=1 then 0) also recovers from the situation.

Change-Id: If842215d4b8f6d795efde1cb40cc52a01841bfa9
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>