iommu: omap: add device OFF support through runtime pm framework
The MMU registers for the remote processors lose their context
in Open Switch Retention (OSWR) or device OFF modes. Hence, the
context of the iommu needs to be saved before it is put to these
lower power state (OSWR/OFF) and restored before it is powered
up to ON again.

Device OFF support for omap iommu domain is implemented through
the .domain_activate and .domain_idle iommu ops. These in turn
call pm_runtime_get_sync and pm_runtime_put on the iommu device.

The PM runtime_suspend and runtime_resume callbacks are registered
with the driver. The IOMMU context is saved through the arch-specific
save context function in the pm_runtime_suspend callback function,
and restored through the existing arch-specific enable function in
the pm_runtime_resume callback.

NOTE: The current implementation is based on the assumption that
are no locked TLBs in use. Logic for saving and restoring locked
TLBs will be added in a future patch.

Change-Id: I9a4f9bed668bd2bc7c74504e604a3faf0b5fed09
Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>