TI-TFW: omap-gpu-governor: hot-spot computation
Based on the characterisation work done by platform enablement
team, there is no need to take pcb sensor input to compute the
hot-spot temperature for GPU domain. Hence before computing the
hot-spot the domain name is checked and if it is for GPU domain
then hot-spot temperature is computed without pcb sensor correction.

Change-Id: Id2c4220c3b12feaf4ae7b9f10e966e0e8a781b7f
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>