ARM: OMAP5: PM: MPU: Update RM CPU0/CPU1 context
Update the register offset for RM_CPUi_CONTEXT.
In OMAP5 RM_CPUi_CPUi_CONTEXT is at offset 0x24 and
in OMAP4 RM_PDA_CPUi_CONTEXT is at offset 0x8.
This patch takes care of addressing this offset change.
The Base addresses and indices are correct, so
in the interest of performance we do not add
runtime checks for capturing the same values.

Acked-by: Vishwanath BS <vishwanath.bs@ti.com>
Signed-off-by: Madhusudan C M <c-madhu@ti.com>