tty: serial: omap-serial: fix irq handling
There are a few problems in serial_omap_irq:

The interrupt type from IIR register is read incorrectly.  UART
interrupts are an integer value masked out of the IIR register,
not a bitmask like any sensible hardware would use.  Convert
the mask operations to mask-and-compare.  Masking iir with
UART_IIR_RLSI was previously matching RDI, RLSI, RX timeout,
and (incorrectly) TX interrupts.  Expand the original mask test
to testing for each of the 3 RX interrupts.

If the interrupt type is UART_IIR_THRI, but the LSR THRE bit
is not set, the interrupt handler doesn't do anything to clear
the THRI interrupt.  Return IRQ_NONE in this case, so the irq
core can detect an interrupt storm and disable the interrupt
if necessary.

[ruchika@ti.com:ported from 3.0 to 3.4 kernel
 commit-id on K3.0:Change-Id:138d34db27e85be07a7a8adf4b2407b7b2ec0390]

Change-Id: I6ab3381ccfe0182961b159a7e9451c3cc466ccda
Signed-off-by: Colin Cross <ccross@android.com>
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>