GPIO: OMAP: support an module specific reset before resume
Certain SoCs are plagued with GPIO module level issues which require
reset of the module when it's context is lost OR if device level
(as against module power domain level) low power state was
achieved. This is very specific to the SoC and the nature
of the specific bug that needs to be workaround. However,
at this point, loss of context may not be sufficient trigger
to reset the module (example an GPIO in wakeup domain still
impacted by a Device level OFF mode bug).

In these cases, support a SoC architecture specific reset functionality
which will verify and decide to reset module or not. This action
is mandatory to be done prior to context restore (as it makes no
sense to restore a module and reset the module).

So, introduce an additional platform level API which can
be hooked and handled by platform code as needed to reset the module
in appropriate manner corresponding to requirement as and when
required.

Change-Id: Ifc53a699958c3a3845a11785be94ae2ebb6647c0
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>
Signed-off-by: Pradeep Gurumath <pradeepgurumath@ti.com>